+++ fpdf/syntax.py
@@ -118,6 +118,14 @@ def create_stream(stream, encryption_handler=None, obj_id=None):
     return "\n".join(["stream", stream, "endstream"])
 
 
+def wrap_in_local_context(draw_commands):
+    """
+    Wrap a series of draw commands (list of strings) in a local context marker, so that changes to
+    draw style only apply to these commands.
+    """
+    return ["q"] + draw_commands + ["Q"]
+
+
 class Raw(str):
     """str subclass signifying raw data to be directly emitted to PDF without transformation."""